Add casts for field-backed parameters and handle ambiguity with multiple constructors

+++ tests/data/enumextension/inner_stubs/expected/Example.java
@@ -3,8 +3,8 @@ import com.example.Stubs.Inner.SubInner;
 
 public enum Example {
     EXISTING(0),
-    A(Inner.A.getParameter(0)),
-    B(Inner.B.getParameter(0)),
-    C(SubInner.C.getParameter(0));
+    A((int) Inner.A.getParameter(0)),
+    B((int) Inner.B.getParameter(0)),
+    C((int) SubInner.C.getParameter(0));
     Example(int parameter) {}
 }
