+++ tests/data/enumextension/stubs/expected/Example.java
@@ -3,8 +3,8 @@ import com.example.Stub2;
 
 public enum Example {
     EXISTING(0),
-    A(RootStub.PROXY.getParameter(0)),
-    B(Stub1.PROXY.getParameter(0)),
-    C(Stub2.PROXY.getParameter(0));
+    A((int) RootStub.PROXY.getParameter(0)),
+    B((int) Stub1.PROXY.getParameter(0)),
+    C((int) Stub2.PROXY.getParameter(0));
     Example(int parameter) {}
 }
